+++ src/Utils/Utils.mm
@@ -6,13 +6,11 @@ cocos2d::CCRect getScreenSafeAreaiOS() {
         UIWindow *keyWindow = [UIApplication sharedApplication].keyWindow;
         UIEdgeInsets insets = keyWindow.safeAreaInsets;
         
-        // Convert from points to Cocos2d-x points (same in this case)
         safeAreaInsets.origin.x = insets.left;
         safeAreaInsets.origin.y = insets.top;
         safeAreaInsets.size.width = insets.right;
         safeAreaInsets.size.height = insets.bottom;
     } else {
-        // Fallback on earlier versions
         safeAreaInsets = cocos2d::CCRect(0, 0, 0, 0);
     }
     return safeAreaInsets;